conftest.py's fixtures are used only in test_update.py
